Made the test code more modular.
