+++ tests/acct_mgt_test.py
@@ -6,7 +6,7 @@
 
  General commandline format:
 
- python3 -m pytest acct-mgt-test.py --amurl [acct_mgt_url] --user [username] --passwd [password]
+ python3 -m pytest amt_test*.py --amurl [acct_mgt_url] --user [username] --passwd [password]
 
  Note, to do this from apple OSX
     1) convert the cert and key to a p12 file
@@ -26,10 +26,10 @@
     curl -u <user>:<password> -kv https://acct-mgt.apps.cnv.massopen.cloud/projects/acct-mgt
 
   -- testing with no authentication:
-     python3 -m pytest acct-mgt-test.py --amurl http://am2.apps.cnv.massopen.cloud
+     python3 -m pytest amt_test*.py --amurl http://acct-mgt..massopen.cloud
 
   -- testing with basic authentication
-     python3 -m pytest acct-mgt-test.py --amurl https://acct-mgt.massopen.cloud --basic "user:pass"
+     python3 -m pytest amt_test*.py --amurl https://acct-mgt.massopen.cloud --basic "user:pass"
 """
 
 import subprocess
@@ -37,7 +37,6 @@ import re
 import time
 import json
 import pprint
-import pytest_check as check
 
 
 def get_pod_status(project, pod_name):
@@ -352,294 +351,3 @@ def ms_user_project_remove_role(
         check=False,
     )
     return compare_results(result, success_pattern)
-
-
-def test_project(acct_mgt_url, auth_opts):
-    """This tests project existence/creation/deletion"""
-    # if oc_resource_exist(
-    #    "project",
-    #    "test-001",
-    #    "test-001[ \t]*test-001[ \t]",
-    #    'Error from server (NotFound): namespaces "test-001" not found',
-    # ):
-    #    print(
-    #        "Error: test_project failed as a project with a name of test-001 exists.  \n" +
-    #        "Please delete first and rerun the tests\n"
-    #    )
-    #    assertTrue(False)
-
-    # test if project doesn't exist
-    check.is_false(
-        ms_check_project(acct_mgt_url, "test-001", auth_opts),
-        "Project exists (test-001)",
-    )
-
-    # test project creation
-    if not oc_resource_exist("project", "Project", "test-001"):
-        check.is_true(
-            ms_create_project(
-                acct_mgt_url, "test-001", r'{"displayName":"test-001"}', auth_opts
-            ),
-            "Project (test-001) not created",
-        )
-        wait_until_done(
-            "oc get project test-001", r"test-001[ \t]+test-001[ \t]+Active"
-        )
-    check.is_true(
-        oc_resource_exist("project", "Project", "test-001"),
-        "Project (test-001) not created",
-    )
-    check.is_true(
-        ms_check_project(acct_mgt_url, "test-001", auth_opts),
-        "project test-001 was not found",
-    )
-
-    # test creation of a second project with the same name
-    if oc_resource_exist("project", "Project", "test-001"):
-        check.is_false(
-            ms_create_project(
-                acct_mgt_url, "test-001", r'{"displayName":"test-001"}', auth_opts
-            ),
-            "Project (test-001) was already created",
-        )
-    check.is_true(
-        oc_resource_exist("project", "Project", "test-001"),
-        "Project test-001 was not found",
-    )
-
-    # test project deletion
-    if oc_resource_exist("project", "Project", "test-001"):
-        check.is_true(
-            ms_delete_project(acct_mgt_url, "test-001", auth_opts),
-            "Unable to delete project (test-001)",
-        )
-        # Wait until test-001 is terminated
-        wait_until_done(
-            "oc get project test-001",
-            r'Error from server (NotFound): namespaces "test-001" not found',
-        )
-    check.is_false(
-        oc_resource_exist("project", "Project", "test-001"),
-        "Project test-001 exists and it shouldn't",
-    )
-
-    # test deleting a project that was deleted
-    if not oc_resource_exist("project", "Project", "test-001"):
-        check.is_false(
-            ms_delete_project(acct_mgt_url, "test-001", auth_opts),
-            "shouldn't be able to delete a non-existing project",
-        )
-    check.is_false(
-        oc_resource_exist("project", "Project", "test-001"),
-        "Project test-001 exists and it should not",
-    )
-
-    # these tests are primarily done to ensure that the microserver doesn't crash
-    #    When the "displayName" is not present, or the json doesn't exist, the displayName shall
-    #    default to the project_uuid (first parameter)
-    check.is_true(
-        ms_create_project(
-            acct_mgt_url,
-            "1234-1234-1234-1234",
-            r'{"displayName":"test-001"}',
-            auth_opts,
-        ),
-        "Project (1234-1234-1234-1234) not created",
-    )
-    ms_delete_project(acct_mgt_url, "1234-1234-1234-1234", auth_opts)
-    check.is_true(
-        ms_create_project(
-            acct_mgt_url, "2234-1234-1234-1234", r'{"displaName":"test-001"}', auth_opts
-        ),
-        "Project (2234-1234-1234-1234) not created",
-    )
-    ms_delete_project(acct_mgt_url, "2234-1234-1234-1234", auth_opts)
-    check.is_true(
-        ms_create_project(acct_mgt_url, "3234-1234-1234-1234", r"{}", auth_opts),
-        "Project (3234-1234-1234-1234) not created",
-    )
-    ms_delete_project(acct_mgt_url, "3234-1234-1234-1234", auth_opts)
-    check.is_true(
-        ms_create_project(acct_mgt_url, "4234-1234-1234-1234", None, auth_opts),
-        "Project (4234-1234-1234-1234) not created",
-    )
-    ms_delete_project(acct_mgt_url, "4234-1234-1234-1234", auth_opts)
-
-
-def test_user(acct_mgt_url, auth_opts):
-    """This tests user (and user identity) existence/creation/deletion"""
-    # if oc_resource_exist(
-    #    "user",
-    #    "test01",
-    #    r"test01[ \t]*[a-f0-9\-]*[ \t]*sso_auth:test01",
-    #    r'Error from server (NotFound): users.user.openshift.io "test01" not found',
-    # ):
-    #    print(
-    #        "Error: test_user failed as a user with a name of test01 exists.\n" +
-    #        "Please delete first and rerun the tests\n"
-    #    )
-    #    assertTrue(False)
-
-    check.is_false(
-        ms_check_user(acct_mgt_url, "test01", auth_opts),
-        "User test01 exists but it shouldn't exist at this point",
-    )
-
-    # test user creation
-    # test01      bfd6dab5-11f3-11ea-89a6-fa163e2bb38b      sso_auth:test01
-    if not oc_resource_exist("users", "User", "test01"):
-        check.is_true(
-            ms_create_user(acct_mgt_url, "test01", auth_opts),
-            "unable to create test01",
-        )
-
-    check.is_true(
-        oc_resource_exist("users", "User", "test01"),
-        "user test01 doesn't exist",
-    )
-    check.is_true(
-        ms_check_user(acct_mgt_url, "test01", auth_opts),
-        "User test01 doesn't exist but it should",
-    )
-
-    # test creation of a second user with the same name
-    if oc_resource_exist("users", "User", "test01"):
-        check.is_false(
-            ms_create_user(acct_mgt_url, "test01", auth_opts),
-            "Should have failed to create a second user with the username of test01",
-        )
-    check.is_true(
-        oc_resource_exist("users", "User", "test01"),
-        "user test01 doesn't exist",
-    )
-
-    # test user deletion
-    if oc_resource_exist("users", "User", "test01"):
-        check.is_true(
-            ms_delete_user(acct_mgt_url, "test01", auth_opts),
-            "user test01 deleted",
-        )
-    check.is_false(
-        oc_resource_exist("users", "User", "test01"),
-        "user test01 not found",
-    )
-
-    # test deleting a user that was deleted
-    if not oc_resource_exist("users", "User", "test01"):
-        check.is_false(
-            ms_delete_user(acct_mgt_url, "test01", auth_opts),
-            "shouldn't be able to delete non-existing user test01",
-        )
-    check.is_false(
-        oc_resource_exist("users", "User", "test01"),
-        "user test01 not found",
-    )
-    check.is_false(
-        ms_check_user(acct_mgt_url, "test01", auth_opts),
-        "User test01 exists but it shouldn't exist at this point",
-    )
-
-
-def test_project_user_role(acct_mgt_url, auth_opts):
-    """This tests project role existence/creation/update/deletion"""
-    # Create a project
-    if not oc_resource_exist("project", "Project", "test-002"):
-        check.is_true(
-            ms_create_project(
-                acct_mgt_url, "test-002", '{"displayName":"test-002"}', auth_opts
-            ),
-            "Project (test-002) was unable to be created",
-        )
-    check.is_true(
-        oc_resource_exist("project", "Project", "test-002"),
-        "Project (test-002) does not exist",
-    )
-
-    # Create some users test02 - test-05
-    for user_number in range(2, 6):
-        if not oc_resource_exist("users", "User", "test0" + str(user_number)):
-            check.is_true(
-                ms_create_user(acct_mgt_url, "test0" + str(user_number), auth_opts),
-                "Unable to create user " + "test0" + str(user_number),
-            )
-        check.is_true(
-            oc_resource_exist("users", "User", "test0" + str(user_number)),
-            "user test0" + str(user_number) + " not found",
-        )
-
-    # now bind an admin role to the user
-    role_info = {"user_name": "test02", "project_name": "test-002", "role": "admin"}
-    check.is_false(
-        ms_user_project_get_role(
-            acct_mgt_url,
-            role_info,
-            r'{"msg": "user role exists \(test-002,test02,admin\)"}',
-            auth_opts,
-        )
-    )
-    check.is_true(
-        ms_user_project_add_role(
-            acct_mgt_url,
-            role_info,
-            r'{"msg": "rolebinding created \(test02,test-002,admin\)"}',
-            auth_opts,
-        ),
-        "Role unable to be added",
-    )
-
-    # should write a oc command to check this, but for rolebindings
-    # this cannot be done in the same way as for users and projects
-    # for now, rely on the microserver.
-    check.is_true(
-        ms_user_project_get_role(
-            acct_mgt_url,
-            role_info,
-            r'{"msg": "user role exists \(test-002,test02,admin\)"}',
-            auth_opts,
-        )
-    )
-
-    check.is_true(
-        ms_user_project_add_role(
-            acct_mgt_url,
-            role_info,
-            r'{"msg": "rolebinding already exists - unable to add \(test02,test-002,admin\)"}',
-            auth_opts,
-        ),
-        "Added the same role to a user failed as it should",
-    )
-
-    check.is_true(
-        ms_user_project_remove_role(
-            acct_mgt_url,
-            role_info,
-            r'{"msg": "removed role from user on project"}',
-            auth_opts,
-        ),
-        "Removed rolebinding successful",
-    )
-    # Should write an oc command to check if a role was added to a user for a project.
-    # however this is not easy based on the current way this is reported by oc so for
-    # now just use the microserver
-    check.is_true(
-        ms_user_project_remove_role(
-            acct_mgt_url,
-            role_info,
-            r'{"msg": "rolebinding does not exist - unable to delete \(test02,test-002,admin\)"}',
-            auth_opts,
-        ),
-        "Unable to remove non-existing rolebinding",
-    )
-
-    # Clean up by removing the users and project (test-002)
-    check.is_true(
-        ms_delete_project(acct_mgt_url, "test-002", auth_opts) is True,
-        "project (test-002) deleted",
-    )
-    for user_number in range(2, 6):
-        if oc_resource_exist("users", "User", "test0" + str(user_number)):
-            check.is_true(
-                ms_delete_user(acct_mgt_url, "test0" + str(user_number), auth_opts)
-                is True,
-                "user " + "test0" + str(user_number) + "unable to be deleted",
-            )
